Fixes issue with retrying a job (issue #4)
The controller method to handle retrying of an individual jobs was not
implemented (retry#update).

It looks like there was an intention for retry#create to handle
retrying of a collection of jobs and an individual job. I personally
like to seperate rest actions on a collection versus an invidual
resource instead of having one action handle both cases. Also, it seems
like to me the "retry" action belongs more with a job resource the same
way the "delete" action currently belongs to a job resource